Rename python package to pysegy and add pyproject
