+++ .github/workflows/main.yml
@@ -47,7 +47,7 @@ jobs:
       - name: Install dependencies
         run: python -m pip install pytest pytest-cov codecov
       - name: Run tests with coverage
-        run: pytest -v --cov --cov-report=xml
+        run: pytest -v --cov=pysegy --cov-report=xml
       - name: Upload coverage
         uses: codecov/codecov-action@v3
         with:
