+++ .gitignore
@@ -2,8 +2,8 @@
 *.jl.*.cov
 *.jl.mem
 __pycache__/
-python/__pycache__/
-python/tests/__pycache__/
+pysegy/__pycache__/
+pysegy/tests/__pycache__/
 .coverage
 coverage.xml
 htmlcov/
