+++ pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pysegy"
+version = "0.1.0"
+description = "Minimal Python implementation of SegyIO"
+authors = [ { name = "SegyIO Developers" } ]
+readme = "README.md"
+requires-python = ">=3.8"
+license = { file = "LICENSE" }
+
+[tool.pytest.ini_options]
+addopts = "-v"
+testpaths = ["pysegy/tests"]
+
