+++ pysegy/read.py
