+++ pysegy/types.py
