+++ pysegy/write.py
